Rename bson formatter to mongo formatter and move tests (#16)
This introduces a breaking change by renaming the BSON formatter to the Mongo formatter. The previous "BSON" name was too generic, as BSON is a data format used by multiple systems. The new "Mongo" naming better reflects that this formatter is specifically designed for MongoDB operations.

This change also sets the foundation for adding additional BSON-based formatters in the future without naming conflicts.